[Ver1.3.2] 배틀 3D 보드 렌더링 버그 수정 — piece.shape base64 인코딩 문제
Go 게이트웨이가 Rust 레퍼리의 piece.shape([][]uint8)를 JSON 재직렬화할 때
uint8=byte로 취급되어 각 행이 base64 문자열로 인코딩됐다. 이 때문에 브라우저의
cellsFor()가 "row.forEach is not a function" 예외를 던지며 3D 보드가 비어 보였다.
Shape 필드를 [][]int로 변경해 숫자 배열로 직렬화되도록 수정 (브라우저 실검증 완료).

+++ backend/go/internal/referee/client.go
@@ -24,11 +24,13 @@ func New(base string) *Client {
 }
 
 type PieceState struct {
-	T     string    `json:"t"`
-	X     int       `json:"x"`
-	Y     int       `json:"y"`
-	Rot   uint8     `json:"rot"`
-	Shape [][]uint8 `json:"shape"`
+	T     string  `json:"t"`
+	X     int     `json:"x"`
+	Y     int     `json:"y"`
+	Rot   uint8   `json:"rot"`
+	// uint8는 []byte로 인식되어 JSON 인코더가 base64로 인코딩한다(→ 프론트 shape가 깨짐).
+	// 숫자 배열로 직렬화되도록 int를 쓴다.
+	Shape [][]int `json:"shape"`
 }
 
 type PlayerState struct {
